+++ gridsome.server.js
@@ -1,10 +1,13 @@
 const VuetifyLoaderPlugin = require('vuetify-loader/lib/plugin')
 const axios = require('axios')
+const format = require('./server/format.js')
+const schema = require('./server/schema.js')
 
 const backendURL = process.env.BACKEND_URL
 const eventsAPI = backendURL + 'api/events'
 const newsAPI = backendURL + 'api/news'
 const contactAPI = backendURL + 'api/contact'
+const calendarAPI = backendURL + 'api/calendar'
 
 module.exports = function (api) {
   api.loadSource((store) => {
@@ -17,7 +20,24 @@ module.exports = function (api) {
     const { data } = await axios.get(eventsAPI + '?beta=' + process.env.BETA)
     const events = actions.addCollection('Event')
     for (const event of data) {
-      events.addNode(event)
+      // convert values
+      const node = JSON.parse(JSON.stringify(event))
+      if (node.costMember) {
+        node.costMember = format.toCurrency(node.costMember)
+      }
+      if (node.costNonMember) {
+        node.costNonMember = format.toCurrency(node.costNonMember)
+      }
+      if (node.durationInMinutes) {
+        node.duration = format.toDuration(node.durationInMinutes)
+        delete node.durationInMinutes
+      }
+      if (node.dates) {
+        node.dates = node.dates.map((date) => format.toDate(date))
+      }
+      // add event
+      events.addNode(node)
+      // create page
       api.createManagedPages(({ createPage }) => {
         if (event.type === 'Fitness') {
           createPage({
@@ -40,39 +60,23 @@ module.exports = function (api) {
     }
   })
 
+  api.loadSource(async (actions) => {
+    const { data } = await axios.get(calendarAPI + '/appointments')
+    const appointments = actions.addCollection('Appointment')
+    for (const appointment of data) {
+      appointments.addNode({
+        sortIndex: appointment.sortIndex,
+        date: format.toDatespan(appointment),
+        time: format.toTimespan(appointment),
+        title: appointment.title,
+        description: appointment.description,
+        link: appointment.link,
+      })
+    }
+  })
+
   api.loadSource(({ addSchemaTypes }) => {
-    addSchemaTypes(`
-      type Event implements Node {
-        id: ID!
-        sheetId: String
-        gid: Int
-        type: String
-        name: String
-        sortIndex: Int
-        visible: Boolean
-        beta: Boolean
-        shortDescription: String
-        description: String
-        image: String
-        light: Boolean
-        dates(
-          format: String
-          locale: String
-        ): [Date]
-        customDate: String
-        durationInMinutes: Int
-        maxSubscribers: Int
-        subscribers: Int
-        costMember: Int
-        costNonMember: Int
-        waitingList: Int
-        maxWaitingList: Int
-        location: String
-        bookingTemplate: String
-        waitingTemplate: String
-        externalOperator: Boolean
-      }
-    `)
+    schema.load(addSchemaTypes)
   })
 
   api.chainWebpack((config, { isServer }) => {